rename to `minify-public-json-files`
