+++ src/plugins/vite/vite-minify-public-json-files.ts
